refactor provider package hierarchy
