+++ client/src/main/kotlin/io/titandata/remote/engine/EngineParameters.kt
@@ -2,7 +2,9 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.models
+package io.titandata.remote.engine
+
+import io.titandata.models.RemoteParameters
 
 data class EngineParameters(
     override var provider: String = "engine",
