+++ client/src/main/kotlin/io/titandata/remote/nop/NopParameters.kt
@@ -2,7 +2,9 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.models
+package io.titandata.remote.nop
+
+import io.titandata.models.RemoteParameters
 
 data class NopParameters(
     override var provider: String = "nop",
