+++ client/src/main/kotlin/io/titandata/remote/nop/NopRemote.kt
@@ -2,7 +2,9 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.models
+package io.titandata.remote.nop
+
+import io.titandata.models.Remote
 
 data class NopRemote(
     override var provider: String = "nop",
