+++ client/src/main/kotlin/io/titandata/remote/s3/S3Remote.kt
@@ -2,7 +2,9 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.models
+package io.titandata.remote.s3
+
+import io.titandata.models.Remote
 
 data class S3Remote(
     override var provider: String = "s3",
