+++ client/src/main/kotlin/io/titandata/remote/ssh/SshParameters.kt
@@ -2,7 +2,9 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.models
+package io.titandata.remote.ssh
+
+import io.titandata.models.RemoteParameters
 
 data class SshParameters(
     override var provider: String = "ssh",
