+++ client/src/main/kotlin/io/titandata/remote/ssh/SshRemoteUtil.kt
@@ -2,13 +2,10 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.serialization.remote
+package io.titandata.remote.ssh
 
 import io.titandata.models.Remote
 import io.titandata.models.RemoteParameters
-import io.titandata.models.SshParameters
-import io.titandata.models.SshRemote
-import io.titandata.serialization.RemoteUtil
 import io.titandata.serialization.RemoteUtilProvider
 import java.io.File
 import java.net.URI
