+++ server/src/test/kotlin/io/titandata/remote/s3/S3RemoteTest.kt
@@ -2,7 +2,7 @@
  * Copyright The Titan Project Contributors.
  */
 
-package io.titandata.serialization
+package io.titandata.remote.s3
 
 import com.google.gson.GsonBuilder
 import io.kotlintest.extensions.system.OverrideMode
@@ -13,8 +13,8 @@ import io.kotlintest.shouldThrow
 import io.kotlintest.specs.StringSpec
 import io.titandata.models.Remote
 import io.titandata.models.RemoteParameters
-import io.titandata.models.S3Parameters
-import io.titandata.models.S3Remote
+import io.titandata.serialization.ModelTypeAdapters
+import io.titandata.serialization.RemoteUtil
 
 class S3RemoteTest : StringSpec() {
